Shoulders gets no "extra" rejuvenation healing
In most cases all it accomplished was moving the GCD used to apply the
rejuvenation and did not actually increase the healing done.

This gives a more honest valuation in all scenarios except full on spam
for over 18s in preparation for a huge damage event. If such a scenario
is revelant use your own judgement to determine wether the shoulders
are good or not.